调整config，删除不必要的代码

+++ config/controller_conf.go
@@ -2,7 +2,7 @@ package config
 
 type Controller struct {
 	BackendListenAddr string `yaml:"BackendListenAddr"`
-	BackendListenPort int64  `yaml:"BackendListenPort"`
+	BackendListenPort uint16 `yaml:"BackendListenPort"`
 	User              string `yaml:"User"`
 	Password          string `yaml:"Password"`
 }
